Move from AlbumServices to ArtistServices
