+++ YouTubeMusicAPI/Services/Artists/ArtistService.cs
@@ -2,6 +2,7 @@
 using YouTubeMusicAPI.Http;
 using YouTubeMusicAPI.Json;
 using YouTubeMusicAPI.Pagination;
+using YouTubeMusicAPI.Services.Albums;
 using YouTubeMusicAPI.Services.Search;
 using YouTubeMusicAPI.Utils;
 
@@ -81,5 +82,104 @@ public sealed class ArtistService
 
         ArtistInfo artist = ArtistInfo.Parse(root);
         return artist;
+    }
+
+    /// <summary>
+    /// Gets the albums of an artist
+    /// </summary>
+    /// <param name="browseId">The browse id of artist</param>
+    /// <param name="params">The browse params, will decide whether to return albums or singles EPs</param>
+    /// <param name="sortingOrder">The sorting order of the returned albums</param>
+    /// <param name="cancellationToken">The token to cancel this task.</param>
+    /// <returns>A list of the <see cref="ArtistAlbum"/> containing the information</returns>
+    /// <exception cref="HttpRequestException">Occurs when the HTTP request fails.</exception>
+    /// <exception cref="OperationCanceledException">Occurs when this task was canceled.</exception>
+    public async Task<ArtistAlbums> GetAlbumsAsync(
+        string browseId,
+        string @params,
+        AlbumSortingOrder sortingOrder = AlbumSortingOrder.Default,
+        CancellationToken cancellationToken = default)
+    {
+        Ensure.NotNullOrEmpty(browseId, nameof(browseId));
+
+        KeyValuePair<string, object?>[] payload =
+        [
+            new("browseId", browseId),
+            new("params", @params)
+        ];
+
+        Task<string> MakeRequest() =>
+            client.RequestHandler.PostAsync(Endpoints.Browse, payload, ClientType.WebMusic, cancellationToken);
+
+        const string methodName = $"{nameof(AlbumService)}-{nameof(GetAlbumsAsync)}";
+        var logger = client.Logger;
+
+        string response = await MakeRequest();
+
+        logger?.LogInformation($"[{methodName}] Parsing response...");
+        using IDisposable _ = response.ParseJson(out JElement root);
+
+        ArtistAlbums ParseAlbums() => ArtistAlbums.Parse(root, browseId, @params, sortingOrder);
+
+        if (sortingOrder is AlbumSortingOrder.Default)
+        {
+            return ParseAlbums();
+        }
+
+        string? continuationToken = null;
+        var sortingOptions = root.Get("contents")
+            .Get("singleColumnBrowseResultsRenderer")
+            .Get("tabs")
+            .GetAt(0)
+            .Get("tabRenderer")
+            .Get("content")
+            .Get("sectionListRenderer")
+            .Get("header")
+            .Get("musicSideAlignedItemRenderer")
+            .Get("endItems")
+            .GetAt(0)
+            .Get("musicSortFilterButtonRenderer")
+            .Get("menu")
+            .GetMultiSelectMenu()
+            .Get("options")
+            .AsArray()
+            .Or(JArray.Empty);
+
+        var sortingOrderString = sortingOrder.ToString().ToLower();
+
+        foreach (var option in sortingOptions)
+        {
+            var optionText = option.Get("musicMultiSelectMenuItemRenderer")
+                .Get("title").GetFirstRun().GetText()
+                .AsString()
+                .OrThrow();
+            if (optionText.Equals(sortingOrderString, StringComparison.InvariantCultureIgnoreCase))
+            {
+                continuationToken = option
+                    .GetMultiSelectMenuItem()
+                    .Get("selectedCommand")
+                    .Get("commandExecutorCommand")
+                    .Get("commands")
+                    .AsArray().OrThrow()
+                    .Last()
+                    .Get("browseSectionListReloadEndpoint")
+                    .Get("continuation")
+                    .Get("reloadContinuationData")
+                    .Get("continuation")
+                    .AsString();
+                break;
+            }
+        }
+
+        if (continuationToken is null)
+        {
+            return ParseAlbums();
+        }
+
+        logger?.LogInformation($"[{methodName}] Resending request to get sorted albums...");
+        payload = [new("continuation", continuationToken)];
+        response = await MakeRequest();
+        using IDisposable __ = response.ParseJson(out root);
+        return ParseAlbums();
     }
 }
